fix(main): change types

+++ main.py
@@ -1,9 +1,9 @@
 import math
 import functions
 
-a = int(input("input a >>"))
-b = int(input("input b >>"))
-c = int(input("input c >>"))
+a = float(input("input a >>"))
+b = float(input("input b >>"))
+c = float(input("input c >>"))
 
 area = functions.calculate_area(a, b, c)
